feat: message with media; refactor: separate every components in messages.tsx
