+++ components/chat/header/options.tsx
@@ -1,8 +1,6 @@
 import { Button } from "@heroui/button";
-import { Image } from "@heroui/image";
-import { addToast } from "@heroui/toast";
-import { ArrowLeft, EllipsisVertical, Search, Trash2 } from "lucide-react";
-import { useMutation, useQuery } from "convex/react";
+import { EllipsisVertical, Trash2 } from "lucide-react";
+import { useMutation } from "convex/react";
 import {
   Modal,
   ModalContent,
@@ -11,7 +9,6 @@ import {
   ModalFooter,
   useDisclosure,
 } from "@heroui/modal";
-import { useRouter } from "next/navigation";
 
 import {
   DropdownMenu,
@@ -23,94 +20,7 @@ import { useChat } from "@/zustand/chat";
 import { api } from "@/convex/_generated/api";
 import { Id } from "@/convex/_generated/dataModel";
 
-export default function ChatHeader() {
-  const router = useRouter();
-
-  const { activeChat, clearActiveChat } = useChat();
-
-  const currentUser = useQuery(api.users.getCurrentUser);
-  const chatById = useQuery(api.chats.getChatById, {
-    _id: activeChat?._id as Id<"chats">,
-  });
-
-  const interlocutor = chatById?.participants.find(
-    (p) => p?._id !== currentUser?._id,
-  );
-
-  return (
-    <div className={`p-4`}>
-      <div className="flex items-center gap-2">
-        <Button
-          isIconOnly
-          className="md:hidden"
-          radius="full"
-          variant="light"
-          onPress={() => {
-            clearActiveChat();
-            router.back();
-          }}
-        >
-          <ArrowLeft />
-        </Button>
-
-        {/* Avatar/Image */}
-        <Image
-          alt="avatar"
-          draggable={false}
-          height={40}
-          radius="full"
-          src={
-            activeChat?.type === "private"
-              ? interlocutor?.avatarUrl
-              : activeChat?.imageUrl
-          }
-          width={40}
-        />
-
-        {/* Content */}
-        <div className="min-w-0 flex-1">
-          {/* Name */}
-          <h2 className="line-clamp-1 text-small font-bold">
-            {activeChat?.type === "private"
-              ? interlocutor?.name
-              : activeChat?.name}
-          </h2>
-
-          {/* Text */}
-          <p className="overflow-hidden text-ellipsis whitespace-nowrap text-small text-default-500">
-            {activeChat?.type === "private"
-              ? interlocutor?.username
-              : activeChat?.description}
-          </p>
-        </div>
-
-        {/* CTA */}
-        <div className="flex items-end gap-1">
-          <Button
-            disableRipple
-            isIconOnly
-            radius="full"
-            variant="light"
-            onPress={() =>
-              addToast({
-                title: "Search messages",
-                description:
-                  "Search through your messages. This feature is coming soon.",
-                color: "warning",
-              })
-            }
-          >
-            <Search size={20} />
-          </Button>
-
-          <Options />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function Options() {
+export default function Options() {
   const { activeChat, clearActiveChat } = useChat();
   const { isOpen, onOpen, onOpenChange } = useDisclosure();
 
@@ -174,12 +84,12 @@ function Options() {
               </ModalHeader>
               <ModalBody>
                 <p>
-                  Are you sure you want to delete this chat? This action
-                  is irreversible.
+                  Are you sure you want to delete this chat? This action is
+                  irreversible.
                 </p>
                 <p>
-                  This will permanently delete the chat for you and the
-                  other participant.
+                  This will permanently delete the chat for you and the other
+                  participant.
                 </p>
               </ModalBody>
               <ModalFooter>
